Move keypairconverter from config to separate module This is a pre-emptive change to prevent a cyclical dependency when trying to use the key-vault-api module in the keypairconverter as part of the upcoming refactor
