+++ key-pair-converter/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>tessera</artifactId>
+        <groupId>com.quorum.tessera</groupId>
+        <version>0.8-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>key-pair-converter</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>com.quorum.tessera</groupId>
+            <artifactId>encryption-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.quorum.tessera</groupId>
+            <artifactId>config</artifactId>
+        </dependency>
+    </dependencies>
+
+
+</project>
